Added to check if the OS is RHEL-based nad fix create-blueprint.sh

+++ setup-local.sh
@@ -1,29 +1,42 @@
 #! /bin/bash
 
-echo "Installing dependencies for Sealight..."
+echo "Installing dependencies for Beanie Builder..."
 
-# Core build tools
-sudo dnf install -y \
-    qemu-img \
-    qemu-kvm \
-    qemu-system-x86 \
-    libvirt \
-    virt-manager \
-    jq \
-    yq \
-    image-builder \
-    edk2-ovmf \
-    python3-pip \
-    zstd \
-    pykickstart \
-    just
+# Check if the host OS is Linux 
+if [ "$(uname -s)" == "Linux" ]; then
+    # Check if the host OS is Fedora, CentOS, or other Red Hat-based distribution
+    if [ -f /etc/redhat-release ]; then
+        # Core build tools
+        sudo dnf install -y \
+            qemu-img \
+            qemu-kvm \
+            qemu-system-x86 \
+            libvirt \
+            virt-manager \
+            jq \
+            yq \
+            image-builder \
+            edk2-ovmf \
+            python3-pip \
+            zstd \
+            pykickstart \
+            just
+        
+        # Install mise
+        sudo dnf copr enable jdxcode/mise
+        sudo dnf install mise -y
+    
+    else
+        echo "Error: Beanie Builder is only supported on Fedora, CentOS, or other Red Hat-based distribution"
+        exit 1
+    fi
+else
+    echo "Error: Beanie Builder is only supported on Linux"
+    exit 1
+fi
 
 # Install tomlq
 pip install tomlq
-
-# Install mise
-sudo dnf copr enable jdxcode/mise
-sudo dnf install mise -y
 
 echo ""
 echo "Setup complete! You can now build Linux OS Images."
